Batch files added to activate venv; threaded scripts added
Threaded versions of the Barnes-Hut and Pairwise simulation added.

Files with the prefix "activate_venv" are now called when the simulation scripts are launched via main.py. This is the only way the simulation scripts will work, if the files are launched manually rather than through main.py, the venv will not be activated and the script will fail. However, navigating to venv/Scripts/activate.bat and then launched the simulation scripts will allow them to work without main.py, however this is a longer method and hence is not advised.